test: UserCrudController -> UserDetailController에 대한 테스트로 수정
