--- app/src/test/java/com/codesoom/assignment/controllers/UserCrudControllerTest.java
@@ -1,68 +1,0 @@
-package com.codesoom.assignment.controllers;
-
-import com.codesoom.assignment.application.interfaces.UserShowService;
-import com.codesoom.assignment.domain.entities.User;
-import com.fasterxml.jackson.databind.ObjectMapper;
-import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Nested;
-import org.junit.jupiter.api.Test;
-import org.mockito.Mock;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.test.web.servlet.MockMvc;
-
-import java.util.List;
-
-import static org.hamcrest.core.StringContains.containsString;
-import static org.mockito.BDDMockito.given;
-import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.*;
-import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.content;
-import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
-
-@SpringBootTest
-@AutoConfigureMockMvc
-@DisplayName("UserCrudController")
-class UserCrudControllerTest {
-    @Mock
-    private UserShowService service;
-    @Autowired
-    private MockMvc mockMvc;
-    @Autowired
-    private ObjectMapper objectMapper;
-
-    private final Long USER_ID = 1L;
-    private final Long USER_ID_NOT_EXISTING = 10L;
-    private final String USER_NAME = "Test User";
-    private final String USER_EMAIL = "hello@gmail.com";
-    private final String USER_PASSWORD = "yahOo~!@12345";
-    private User user;
-
-    @BeforeEach
-    void setUp() {
-        user = User.builder()
-                .id(USER_ID)
-                .name(USER_NAME)
-                .email(USER_EMAIL)
-                .password(USER_PASSWORD)
-                .build();
-    }
-
-    @Nested
-    @DisplayName("list 메소드는")
-    class Describe_list {
-        @BeforeEach
-        void setUp() {
-            given(service.showAll()).willReturn(List.of(user));
-        }
-
-        @Test
-        @DisplayName("HTTP Status Code 200 OK 응답한다")
-        void it_responds_with_200_ok() throws Exception {
-            mockMvc.perform(get("/users"))
-                    .andExpect(status().isOk())
-                    .andExpect(content().string(containsString(USER_NAME)));
-        }
-    }
-}
