+++ app/src/test/java/com/codesoom/assignment/controllers/UserDetailControllerTest.java
@@ -1,0 +1,88 @@
+package com.codesoom.assignment.controllers;
+
+import com.codesoom.assignment.application.exceptions.ProductNotFoundException;
+import com.codesoom.assignment.application.UserCrudService;
+import com.codesoom.assignment.application.exceptions.UserNotFoundException;
+import com.codesoom.assignment.domain.entities.User;
+import com.fasterxml.jackson.databind.ObjectMapper;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Nested;
+import org.junit.jupiter.api.Test;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.boot.test.mock.mockito.MockBean;
+import org.springframework.test.web.servlet.MockMvc;
+
+import static org.mockito.BDDMockito.given;
+import static org.mockito.Mockito.mock;
+import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
+import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
+
+@SpringBootTest
+@AutoConfigureMockMvc
+@DisplayName("UserDetailController")
+class UserDetailControllerTest {
+    @MockBean
+    private UserCrudService service;
+    @Autowired
+    private MockMvc mockMvc;
+    @Autowired
+    private ObjectMapper objectMapper;
+
+    private final Long USER_ID = 1L;
+    private final Long USER_ID_NOT_EXISTING = 10L;
+    private final String USER_NAME = "Test User";
+    private final String USER_EMAIL = "hello@gmail.com";
+    private final String USER_PASSWORD = "yahOo~!@12345";
+    private User user;
+
+    @BeforeEach
+    void setUp() {
+        user = User.builder()
+                .id(USER_ID)
+                .name(USER_NAME)
+                .email(USER_EMAIL)
+                .password(USER_PASSWORD)
+                .build();
+    }
+
+    @Nested
+    @DisplayName("detail 메소드는")
+    class Describe_detail {
+        @Nested
+        @DisplayName("유효한 매개변수를 전달 받는다면")
+        class Context_with_valid_param {
+            @BeforeEach
+            void setUp() {
+                given(service.showById(USER_ID)).willReturn(user);
+            }
+
+            @Test
+            @DisplayName("HTTP Status Code 200 OK 응답한다")
+            void it_responds_with_200_ok() throws Exception {
+                mockMvc.perform(get("/users/" + USER_ID))
+                        .andExpect(status().isOk());
+            }
+        }
+
+        @Nested
+        @DisplayName("만약 존재하지 않는 ID를 매개변수로 전달 받는다면")
+        class Context_without_existing_toy {
+            @BeforeEach
+            void setUp() {
+                given(service.showById(USER_ID_NOT_EXISTING))
+                        .willThrow(new UserNotFoundException(USER_ID_NOT_EXISTING));
+            }
+
+            @Test
+            @DisplayName("HTTP Status Code 404 NOT FOUND 응답한다")
+            void it_responds_with_404() throws Exception {
+                mockMvc.perform(get("/users/" + USER_ID_NOT_EXISTING))
+                        .andExpect(status().isNotFound());
+            }
+
+        }
+    }
+}
